feat: 이슈 필터링
- hashMap을 통한 필터링 조건 관리

- 커스텀 뷰를 통해 조건 뷰 재활용

- 프래그먼트간 데이터 전달을 위해 arguments 이용

- spinner을 통해 조건 리스트 선택

+++ app/src/main/java/com/example/it/issuetracker/presentation/main/issue/IssueViewModel.kt
@@ -6,7 +6,9 @@ import com.example.it.issuetracker.domain.model.Issue
 import com.example.it.issuetracker.domain.repository.IssueTrackerRepository
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
+import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
+import java.util.*
 
 class IssueViewModel(
     private val issueRepository: IssueTrackerRepository,
@@ -15,14 +17,18 @@ class IssueViewModel(
     private val _uiState = MutableStateFlow<IssueUiState>(IssueUiState.UnInitialization)
     val uiState = _uiState.asStateFlow()
 
+    private val _cache = MutableStateFlow<CacheIssue>(CacheIssue(sortedMapOf(), ""))
+    val cache = _cache.asStateFlow()
+
     init {
         getIssues()
     }
 
     private fun getIssues() = viewModelScope.launch {
-        _uiState.value = IssueUiState.Loading
+        _uiState.update { IssueUiState.Loading }
         val activeIssues = issueRepository.getIssue().getOrThrow()
-        _uiState.value = IssueUiState.GetIssues(activeIssues)
+        _uiState.update { IssueUiState.GetIssues(activeIssues) }
+        _cache.value = CacheIssue(sortedMapOf(), "")
     }
 
     fun toggleViewType() {
@@ -34,7 +40,8 @@ class IssueViewModel(
                         Mode.EDIT -> issue.copy(viewType = Mode.DEFAULT)
                     }
                 }
-                _uiState.value = IssueUiState.GetIssues(editIssueList)
+                _uiState.update { IssueUiState.GetIssues(editIssueList) }
+                _cache.value = CacheIssue(sortedMapOf(), "")
             }
             else -> {}
         }
@@ -47,26 +54,56 @@ class IssueViewModel(
                     (_uiState.value as IssueUiState.GetIssues).issues.map { issue ->
                         issue.copy(viewType = Mode.DEFAULT)
                     }
-                _uiState.value = IssueUiState.GetIssues(defaultIssueList)
+                _uiState.update { IssueUiState.GetIssues(defaultIssueList) }
+                _cache.value = CacheIssue(sortedMapOf(), "")
+
             }
             else -> {}
         }
     }
 
     fun deleteIssue() = viewModelScope.launch {
-        val filterList =
-            (_uiState.value as IssueUiState.GetIssues).issues.filter { issue -> issue.isChecked }
-        if (isCheckEmpty(filterList)) return@launch
-        val activeIssues = issueRepository.deleteIssue(filterList).getOrThrow()
-        _uiState.value = IssueUiState.GetIssues(activeIssues)
+        when (_uiState.value) {
+            is IssueUiState.GetIssues -> {
+                val filterList =
+                    (_uiState.value as IssueUiState.GetIssues).issues.filter { issue -> issue.isChecked }
+
+                val newList: SortedMap<Int, Issue> = sortedMapOf()
+                (_uiState.value as IssueUiState.GetIssues).issues.mapIndexed { idx, issue ->
+                    if (issue.isChecked) {
+                        newList[idx] = issue
+                    }
+                }
+
+                if (isCheckEmpty(filterList)) return@launch
+                val activeIssues = issueRepository.deleteIssue(filterList).getOrThrow()
+                _uiState.update { IssueUiState.GetIssues(activeIssues) }
+                _cache.value = CacheIssue(newList, "선택한 이슈를 삭제했습니다.")
+            }
+            else -> {}
+        }
     }
 
     fun closeIssue() = viewModelScope.launch {
-        val filterList =
-            (_uiState.value as IssueUiState.GetIssues).issues.filter { issue -> issue.isChecked }
-        if (isCheckEmpty(filterList)) return@launch
-        val activeIssue = issueRepository.closeIssue(filterList).getOrThrow()
-        _uiState.value = IssueUiState.GetIssues(activeIssue)
+        when (_uiState.value) {
+            is IssueUiState.GetIssues -> {
+                val filterList =
+                    (_uiState.value as IssueUiState.GetIssues).issues.filter { issue -> issue.isChecked }
+
+                val newList: SortedMap<Int, Issue> = sortedMapOf()
+                (_uiState.value as IssueUiState.GetIssues).issues.mapIndexed { idx, issue ->
+                    if (issue.isChecked) {
+                        newList[idx] = issue
+                    }
+                }
+
+                if (isCheckEmpty(filterList)) return@launch
+                val activeIssue = issueRepository.closeIssue(filterList).getOrThrow()
+                _uiState.update { IssueUiState.GetIssues(activeIssue) }
+                _cache.value = CacheIssue(newList, "선택한 이슈를 닫았습니다.")
+            }
+            else -> {}
+        }
     }
 
     private fun isCheckEmpty(filterList: List<Issue>): Boolean {
@@ -75,5 +112,14 @@ class IssueViewModel(
             return true
         }
         return false
+    }
+
+    fun revertIssue(list: SortedMap<Int, Issue>) = viewModelScope.launch {
+        val activeIssues = issueRepository.revertIssue(list).getOrThrow()
+        _uiState.update { IssueUiState.GetIssues(activeIssues) }
+    }
+
+    fun getFilterList(value: List<Issue>) {
+        _uiState.update { IssueUiState.GetIssues(value) }
     }
 }
